move code out of couchforms.utils
