+++ corehq/apps/receiverwrapper/views.py
@@ -14,6 +14,7 @@ from corehq.apps.receiverwrapper.auth import (
     domain_requires_auth,
 )
 from corehq.apps.receiverwrapper.util import get_app_and_build_ids, determine_authtype
+from corehq.form_processor.submission_post import SubmissionPost
 from corehq.form_processor.utils import convert_xform_to_json
 from corehq.util.datadog.utils import count_by_response_code
 import couchforms
@@ -26,7 +27,7 @@ def _process_form(request, domain, app_id, user_id, authenticated,
                   auth_cls=AuthContext):
     instance, attachments = couchforms.get_instance_and_attachment(request)
     app_id, build_id = get_app_and_build_ids(domain, app_id)
-    response = couchforms.SubmissionPost(
+    response = SubmissionPost(
         instance=instance,
         attachments=attachments,
         domain=domain,
