+++ corehq/apps/tzmigration/timezonemigration.py
@@ -8,6 +8,7 @@ from casexml.apps.case.cleanup import rebuild_case_from_actions
 from casexml.apps.case.models import CommCareCase, CommCareCaseAction
 
 from casexml.apps.case.xform import get_case_updates
+from corehq.form_processor.utils.metadata import scrub_meta
 from dimagi.utils.couch.database import iter_docs
 from corehq.apps.tzmigration import set_migration_started, \
     set_migration_complete
@@ -16,7 +17,6 @@ from corehq.form_processor.parsers.ledgers import get_stock_actions
 from corehq.form_processor.utils import convert_xform_to_json, adjust_datetimes
 from couchforms.dbaccessors import get_form_ids_by_type
 from couchforms.models import XFormInstance
-from couchforms.util import scrub_meta
 from corehq.apps.tzmigration import force_phone_timezones_should_be_processed
