+++ corehq/ex-submodules/couchforms/__init__.py
@@ -1,3 +1,3 @@
 from .exceptions import XMLSyntaxError, CouchFormException, UnexpectedDeletedXForm
 from .getters import *
-from .util import SubmissionPost, fetch_and_wrap_form
+from .util import fetch_and_wrap_form
