+++ corehq/ex-submodules/couchforms/exceptions.py
@@ -1,3 +1,6 @@
+from dimagi.utils.mixins import UnicodeMixIn
+
+
 class CouchFormException(Exception):
     """
     A custom exception for the XForms application.
@@ -18,3 +21,15 @@ class UnexpectedDeletedXForm(Exception):
     pass
 
 
+class SubmissionError(Exception, UnicodeMixIn):
+    """
+    When something especially bad goes wrong during a submission, this
+    exception gets raised.
+    """
+
+    def __init__(self, error_log, *args, **kwargs):
+        super(SubmissionError, self).__init__(*args, **kwargs)
+        self.error_log = error_log
+
+    def __str__(self):
+        return str(self.error_log)
