+++ corehq/form_processor/utils/__init__.py
@@ -5,7 +5,6 @@ from .general import (
 from .xform import (
     extract_meta_instance_id,
     extract_meta_user_id,
-    new_xform,
     convert_xform_to_json,
     adjust_datetimes,
 )
