+++ corehq/form_processor/utils/metadata.py
@@ -7,6 +7,54 @@ from dimagi.utils.parsing import json_format_datetime
 from corehq.util.dates import iso_string_to_datetime
 
 
+def scrub_meta(xform):
+    """
+    Cleans up old format metadata to our current standard.
+
+    Does NOT save the doc, but returns whether the doc needs to be saved.
+    """
+    property_map = {'TimeStart': 'timeStart',
+                    'TimeEnd': 'timeEnd',
+                    'chw_id': 'userID',
+                    'DeviceID': 'deviceID',
+                    'uid': 'instanceID'}
+
+    if not hasattr(xform, 'form'):
+        return
+
+    # hack to make sure uppercase meta still ends up in the right place
+    found_old = False
+    if 'Meta' in xform.form:
+        xform.form['meta'] = xform.form['Meta']
+        del xform.form['Meta']
+        found_old = True
+    if 'meta' in xform.form:
+        meta_block = xform.form['meta']
+        # scrub values from 0.9 to 1.0
+        if isinstance(meta_block, list):
+            if isinstance(meta_block[0], dict):
+                # if it's a list of dictionaries, arbitrarily pick the first one
+                # this is a pretty serious error, but it's also recoverable
+                xform.form['meta'] = meta_block = meta_block[0]
+                logging.error((
+                    'form %s contains multiple meta blocks. '
+                    'this is not correct but we picked one abitrarily'
+                ) % xform.get_id)
+            else:
+                # if it's a list of something other than dictionaries.
+                # don't bother scrubbing.
+                logging.error('form %s contains a poorly structured meta block.'
+                              'this might cause data display problems.')
+        if isinstance(meta_block, dict):
+            for key in meta_block:
+                if key in property_map and property_map[key] not in meta_block:
+                    meta_block[property_map[key]] = meta_block[key]
+                    del meta_block[key]
+                    found_old = True
+
+    return found_old
+
+
 def clean_metadata(_meta_block):
     meta_block = copy(dict(_meta_block))
     if not meta_block:
